Initial commit of KFParticle integration in sPHENIX
This package contais v0.2 of the KFParticle sPHENIX integration package. There are a set of instructions at instructions/KFParticle.pdf and a test example at Fun4ALL_G4_Readback.C

The instructions contain a list of to-do's and ideas so please send me an email if you wold like to help out.