Major code changes to KFParticle
I've made some major changes to the code to move towards the package handling intermediate states. Several functions had been split into different parts and conditions added to resuse them for other reconstruction (such as J/psi's).

The daughter track requirements are now written into KFParticle by the user as a pair where std::pair<std::string, int> is the name of the particle and charge respectively. The maximum number of tracks per vertex is still 4.

This has all been tested with D0->Kpi and produces the same results as before.

Next step is to build a function to handle intermediate states using the other functions in the package.