Some updates to KFParticle package
Updated KFParticle so that all declared particles now have their PDG ID as well as their mass in KFParticle_particleList. The PDG ID is then written to the output file so you can compare with the true ID for efficiency studies. The daughter tracks are required to exist in KFParticle_particleList which was always a requirement but now KFParticle will check that these daughters do exist and, if they dont, will print the offending particle string and exit. If the intermediate or mother particles exist in the particle list, their PDG ID will also be written to the nTuple or else the branch will be populated with zeros.

If the mother is not constrained to a production vertex, variables that require this information are no longer declared in the nTuple. This means no primary vertex information, no decay time/length, no flight distance chi2, no IP or IP chi2 and no DIRA.

Also, smoothed off some code according to cppcheck output.