Added a KFParticle container
This update adds in a container for KFParticle objects and adds it to the users DST. Current it is added under the DST branch so you get a unique set of reconstructed particles for each event rather than a whole set of particles with no idea about the rest of the event. Maybe this could be fixed later?

The KFParticle and selected tracks can be saved with a unique node name which will allow us to create separate containers for basics particles such as phis and Ks0

I also got rid of all mentions of std:: and added a KFParticle identify function