[comment]: <> (Please tell us something about this pull request)
[comment]: <> ( What types of changes does your code introduce? Put an `x` in all the boxes that apply: )
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work for users)
- [ ] Requiring change in macros repository (Please provide links to the macros pull request in the last section)
- [x] I am a member of [GitHub organization of sPHENIX Collaboration](https://github.com/orgs/sPHENIX-Collaboration/people) (contact Chris Pinkenburg to join)

[comment]: <> ( What does this PR do? Linking to talk in software meeting encouraged )

This is the initial push of the KFParticle project for event reconstruction at sPHENIX.

The project was initially developed in the analysis repository before this move to core software. clang-format and cppcheck have previously been run so hopefully there isn't too much work to be done after jenkins runs over this. The corresponding software meeting link for this project is https://indico.bnl.gov/event/7445/

KFParticle was developed outside sPHENIX and the base code is not maintained by us. This project is effectively a wrapper for Fun4All and uses KFParticle object to perform a more generic event reconstruction. The main features of the project are:
(1) Variable input sources - The project translates SvtxTrack's and SvtxVertex's into KFParticle objects but the node used can be specified by the user at configuration. If different input styles are needed then a short function to extract and trranslate the style must be produced.
(2) General event reconstruction - The user configures the reconstruction style (all final state tracks, all intermediate states or a mix of the two) and some selctions commonly used in high energy (nuclear) physics analysis. KFParticle then searches your input files for anything matching this selection.
(3) Output written directly to an nTuple or objects placed back on node tree - The user can specify to save their selection to an nTuple with one boolean option. The package then takes care of creating a sensible output file for you, no need to specify branches or names (except to enable truth matching or detector info). Similarly, the user can specify whether the candidates should be appended back to the node tree with one boolean option as an SvtxTrack and/or KFParticle container. The new SvtxTrackMap will also contain the reconstructed particles like J/psi or Lambda_c, interpreted as a track

Example macros can still be found in the analysis repository. A G4 macro for the macros repo will be developed soon.

[comment]: <> ( In case this is a draft PR, e.g. for run checks using Jenkins, please add the Label `preview-do-not-merge` on the right column  )

This is a draft PR to be tested by Jenkins before a full merge and should only exist on my own KFParticle branch

I need to test the MVA calculation in KFParticle. Some algorithm must be trained first so this will probably happen soon after MDC-1 when we have background samples to train on
I need to make a G4 macro so we can use KFParticle in our more general testing/user framework