Added an assert in KFParticle_eventReconstruction.cc
There was a warning (resulting in a failure from Jenkins) about an array potentially having a negative size. The array was based on how many reconstructed particles the user had and thus could never be negative under normal circumstances (undefined behavior on the other hand...). An assert has been added to keep things happy and safe.