Fix to KFParticle DST output
I noticed when generating MDC data that when creating a track map for the loose reconstruction, the tracks used disappear from the original track map rather than being copied from them. I cloned the original track map then made the new container from there which keeps the original tracks in place. This might be related to the crash from jet truth matching which could be trying to access the missing tracks.

I also increased the verbosity level for the DST class as it mostly duplicates the basic reconstruction output and would thus be annoying for most users.