Added true track-vertex DCA in nTuple output
A request was made to calculate the true DCA in the nTuple output. Thisis calcualted from the true track parameters and the VtxPoint, if it exists. If the true VtxPoint does not exist then the true DCA is calculated wrt to the reconstructed vertex and a warning is given.

I also added some new DCA branches in the nTuple: DCA_xy and DCA error to aid in resolution studies. Please note that the DCA_xy error is not calculated. I also changed the units of the decay time from ctau to ps.

Lastly, the magnetic field was set to negative 1.5T. I changed this to positive 1.5T which matches our magnetic field map plot but I should check that there isn't a flip in the sign somewhere.