Adding true DCA calculations to KFParticle output (nTuple only) (#1058)
* Fix to KFParticle DST output

I noticed when generating MDC data that when creating a track map for the loose reconstruction, the tracks used disappear from the original track map rather than being copied from them. I cloned the original track map then made the new container from there which keeps the original tracks in place. This might be related to the crash from jet truth matching which could be trying to access the missing tracks.

I also increased the verbosity level for the DST class as it mostly duplicates the basic reconstruction output and would thus be annoying for most users.

* clang didn't like my logic statement for verbosity

* clang still didn't like my logic

* Fix KFParticle bug

Found a solution to the KFParticle missing track bag that uses the least memory. There was a track reset call on the reconstructed tracks. This has been removed and the map is no longer cloned.

Also changed the verbosity code to use what is inherited from Fun4AllBase.h

* Added true track-vertex DCA in nTuple output

A request was made to calculate the true DCA in the nTuple output. Thisis calcualted from the true track parameters and the VtxPoint, if it exists. If the true VtxPoint does not exist then the true DCA is calculated wrt to the reconstructed vertex and a warning is given.

I also added some new DCA branches in the nTuple: DCA_xy and DCA error to aid in resolution studies. Please note that the DCA_xy error is not calculated. I also changed the units of the decay time from ctau to ps.

Lastly, the magnetic field was set to negative 1.5T. I changed this to positive 1.5T which matches our magnetic field map plot but I should check that there isn't a flip in the sign somewhere.

* Changed magnetic field from +1.5T to -1.4T

* Change to KFParticle truth matching

I made a small change that the true DCA calcualtion will only be performed if the vertex constraint is applied. Other wise the matching will be done against the mother track and return a non-physical measurement.