KFParticle: 2 fixes (ghosts and resonance/track bug) (#1085)
* KFParticle fix for ghost tracks

Occasionally we can reconstruct tracks from noise hits. When we try to truth match these tracks we get a null pointer and KFParticle crashes. I've added a check for this null pointer and set the truth variables to 0.

* KFParticle: bug fix for events with tracks and resonances

I found a bug in KFParticle reconstruction that afects analysis where a mother decays to a resonance and reconstructed tracks AND the resonance mass was not contrained (such as B+ -> D0 pi+ where the D0 mass was not fixed to the PDG mass). This would cause all the tracks to have no mass and no PID (see https://indico.bnl.gov/event/10855/contributions/45863/attachments/32826/52483/MDC1_C_Dean_20210222.pdf). This PR now fixes that and a small issue where tracks used by resonance weren't properly removed from the good track list.

I also added a few more particle names to the particleList class