Update Makefile.am
The TrackEvaluation.cc crashes the build with a warning. The new container class does not have a virtual base class. After last weeks disaster which causes a ton of work we will just not allow this anymore. If it can be saved on a DST it has to have a virtual base class