Add additional track sources to event evaluator
This enables storing the tracking output from multiple sources of tracks
in the event evaluator (say, running the tracking twice with different
parameters). The tracks are stored in one collection, and then
differentiated by a source ID. The user then needs to differentiate the
sources. However, if there is only one collection of tracks, then the
source can be ignored