Fix distortion corrections for regions outside of the real TPC volume
previously, the corrections are instantiated as dz=z, dr=r, dphi=phi in order to let 'auto' give the new objects the same type as the original.  I added lines to set these to zero, so that when the histogram interpolation fails, it returns zero rather than corrections exactly equal to the original position.